Add and change: Transportation page
The general table is outside the form, and inside it only the event and address fields are displayed. I added a map to automatically calculate the distance between the venue's address and the selected one, filling out that field of the form. If no addresses are available, a new one can be added from the same form.